Update acl code and unit test
